huazie_调整失效时间类型【long到int】_MOD

+++ flea-frame-cache/src/main/java/com/huazie/frame/cache/memcached/MemCachedSpringCacheManager.java
@@ -46,7 +46,7 @@ public class MemCachedSpringCacheManager extends AbstractSpringCacheManager {
     }
 
     @Override
-    protected AbstractSpringCache newCache(String name, long expiry) {
+    protected AbstractSpringCache newCache(String name, int expiry) {
         return new MemCachedSpringCache(name, expiry, memCachedClient);
     }
